Implemented elo logic and edit rounds calcuation

+++ logic/profileLogic.py
@@ -16,7 +16,7 @@ class Profile(db.Model):
     date_added = db.Column(db.DateTime, nullable=True)
 
     # Game Stats
-    elo = db.Column(db.Integer, default=1000)
+    elo = db.Column(db.Float, default=1000.0)
 
     winner_percentage = db.Column(db.Integer, default=0)
